moving sqlitePragmaString function to SqliteRepository file. Migrated from github.com/glebarez/sqlite to gorm.io/driver/sqlite (which uses github.com/mattn/go-sqlite3)
Override github.com/mattn/go-sqlite3 with forked version from @jgiannuzzi which supports Encryption at rest.
See https://github.com/mattn/go-sqlite3/pull/1109
See https://github.com/fastenhealth/fasten-onprem/issues/236

Added documentation for how to open an encrypted sqlite database in IntelliJ - CONTRIBUTING.md update
